# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cultivar Collection Management API",
    "description": "Advanced cannabis cultivar tracking and analytics platform",
    "version": "2.0.0",
    "x-generated": {
      "timestamp": "2025-11-01T19:55:58.334401",
      "source": "app.fastapi_app",
      "generator": "scripts/generate_openapi.py"
    }
  },
  "paths": {
...
    "/social/share": {
      "post": {
        "tags": [
          "Social",
          "social"
        ],
        "summary": "Social Share",
        "description": "Handle social media sharing.",
        "operationId": "social_share_social_share_post",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Social media platform",
              "title": "Platform"
            },
            "description": "Social media platform"
          },
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL to share",
              "title": "Url"
            },
            "description": "URL to share"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Share title",
              "default": "Check out CultivAR!",
              "title": "Title"
            },
            "description": "Share title"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Share description",
              "title": "Description"
            },
            "description": "Share description"
          },
          {
            "name": "hashtags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated hashtags",
              "title": "Hashtags"
            },
            "description": "Comma-separated hashtags"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/share/blog/{slug}": {
      "get": {
        "tags": [
          "Social",
          "social"
        ],
        "summary": "Social Share Blog",
        "description": "Generate social sharing URLs for a blog post.",
        "operationId": "social_share_blog_social_share_blog__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Specific platform to share to",
              "title": "Platform"
            },
            "description": "Specific platform to share to"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Custom title",
              "title": "Title"
            },
            "description": "Custom title"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Custom description",
              "title": "Description"
            },
            "description": "Custom description"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/social/widgets/follow-buttons": {
      "get": {
        "tags": [
          "Social",
          "social"
        ],
        "summary": "Social Follow Buttons",
        "description": "Generate follow buttons for embedding.",
        "operationId": "social_follow_buttons_social_widgets_follow_buttons_get",
        "parameters": [
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Button style: horizontal, vertical, compact",
              "default": "horizontal",
              "title": "Style"
            },
            "description": "Button style: horizontal, vertical, compact"
          },
          {
            "name": "platforms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated platform list",
              "default": "twitter,facebook,instagram",
              "title": "Platforms"
            },
            "description": "Comma-separated platform list"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/widgets/share-buttons": {
      "get": {
        "tags": [
          "Social",
          "social"
        ],
        "summary": "Social Share Buttons",
        "description": "Generate share buttons for embedding.",
        "operationId": "social_share_buttons_social_widgets_share_buttons_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "URL to share",
              "title": "Url"
            },
            "description": "URL to share"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Share title",
              "default": "CultivAR - Professional Cannabis Grow Management",
              "title": "Title"
            },
            "description": "Share title"
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Button style",
              "default": "horizontal",
              "title": "Style"
            },
            "description": "Button style"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}